+++ src/sdk/CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Request retry mechanism, will handle core http and Graph batch requests #21 [jansenbe - Bert Jansen]
 - Tweaking and updating of the writing tests documentation #65 [pkbullock - Paul Bullock]
 - Paging support for Graph/Rest #3 [jansenbe - Bert Jansen]
+- Added Clone method on PnPContext #49 [jansenbe - Bert Jansen]
 
 ### Changed